requirements.txt
gunicorn
djangorestframework
psycopg2
django-heroku
pillow
django-imagekit
Django-Verify-Email
django.utils.six
django-mptt
django-six
bootstrap4
bootstrap_datepicker_plus
